Add validation option name typing
Sometimes it is useful to get the validation option boolean return value inside the script rather than inside the template. One case is to integrate vuelidate validations into vuetify text-field error messages. The error-messages prop accepts an array of strings, so the best way to integrate vuelidate into it is to make a method that returns this array.